remove warmup process
